Move Phrase to `app.api.entity`
